test(agnocast_kmod): move the discovery agent kunit cases out of add_process
