+++ agnocast_kmod/agnocast_kunit/Makefile.inc
@@ -15,6 +15,7 @@ KUNIT_TEST_OBJS := \
   agnocast_kunit_add_bridge.o \
   agnocast_kunit_remove_bridge.o \
   agnocast_kunit_domain_bridge.o \
+  agnocast_kunit_discovery_agent.o \
   agnocast_kunit_set_ros2_subscriber_num.o \
   agnocast_kunit_set_ros2_publisher_num.o \
   agnocast_kunit_do_exit.o \
