+++ agnocast_kmod/agnocast_kunit_main.c
@@ -6,6 +6,7 @@
 #include "agnocast_kunit/agnocast_kunit_add_subscriber.h"
 #include "agnocast_kunit/agnocast_kunit_bridge_shutdown.h"
 #include "agnocast_kunit/agnocast_kunit_check_and_request_bridge_shutdown.h"
+#include "agnocast_kunit/agnocast_kunit_discovery_agent.h"
 #include "agnocast_kunit/agnocast_kunit_do_exit.h"
 #include "agnocast_kunit/agnocast_kunit_domain_bridge.h"
 #include "agnocast_kunit/agnocast_kunit_get_node_publisher_topics.h"
@@ -43,6 +44,7 @@ struct kunit_case agnocast_test_cases[] = {
   TEST_CASES_PUBLISH_MSG,
   TEST_CASES_TAKE_MSG,
   TEST_CASES_ADD_PROCESS,
+  TEST_CASES_DISCOVERY_AGENT,
   TEST_CASES_BRIDGE_SHUTDOWN,
   TEST_CASES_CHECK_AND_REQUEST_BRIDGE_SHUTDOWN,
   TEST_CASES_GET_SUBSCRIBER_NUM,
